requirements.txt
pyteal==0.9.1
py-algorand-sdk==1.8.0
mypy==0.910
pytest
black==22.6.0
websocket-client
requests
jsonpath-ng